train model/ultralytics-8.3.163/requirements.txt
ultralytics==8.0.114
pandas
opencv-python
numpy
scipy
easyocr
filterpy
streamlit
numpy
Pillow
opencv-python-headless
